// src/app/features/company/models/company.model.ts
export interface Company {
    id: string;
    name: string;
    shortName: string;
    addressLine1: string;
    addressLine2: string;
    zipCode: string;
    emailId: string;
    phoneNo1: string;
    phoneNo2: string;
    vatNo: string;
    crNo: string;
    logo: string;
    currencyId : number;
    currency : string;
}